Address further PR comments
